work: gearboxes spawn respective bronzes

+++ src/main/java/corundum/rubinated_nether/content/TarnishStage.java
@@ -17,11 +17,11 @@ import net.minecraft.world.level.block.Block;
 import java.util.function.IntFunction;
 
 public enum TarnishStage {
-    UNAFFECTED((byte) 0,"unaffected", 0.25D, 20.0, 5.0, 0.0, 1200),
-    DISCOLORED((byte) 1, "discolored", 0.20D, 20.0, 10.0, 8.0, 1600),
-    CORRODED((byte) 2, "corroded", 0.15D, 20.0, 15.0, 16.0, 2000),
-    TARNISHED((byte) 3, "tarnished", 0.10D, 20.0, 10.0, 20.0, 2400),
-    CRYSTALLIZED((byte) 4, "crystallized", 0.32D, 8.0, 4.0, 0.0, -1);
+    UNAFFECTED((int) 0,"unaffected", 0.25D, 20.0, 5.0, 0.0, 1200),
+    DISCOLORED((int) 1, "discolored", 0.20D, 20.0, 10.0, 8.0, 1600),
+    CORRODED((int) 2, "corroded", 0.15D, 20.0, 15.0, 16.0, 2000),
+    TARNISHED((int) 3, "tarnished", 0.10D, 20.0, 10.0, 20.0, 2400),
+    CRYSTALLIZED((int) 4, "crystallized", 0.32D, 8.0, 4.0, 0.0, -1);
 
     public static final IntFunction<TarnishStage> BY_ID = ByIdMap.continuous(TarnishStage::getId, values(), ByIdMap.OutOfBoundsStrategy.ZERO);
 
@@ -38,7 +38,7 @@ public enum TarnishStage {
         }
     };
 
-    private final byte id;
+    private final int id;
     private final String name;
     final double speed;
     final double health;
@@ -46,7 +46,7 @@ public enum TarnishStage {
     final double armor;
     final int tarnishDuration;
 
-    TarnishStage(byte id, String name, double speed, double health, double attack, double armor, int tarnishDuration) {
+    TarnishStage(int id, String name, double speed, double health, double attack, double armor, int tarnishDuration) {
         this.id = id;
         this.name = name;
         this.speed = speed;
@@ -56,7 +56,7 @@ public enum TarnishStage {
         this.tarnishDuration = tarnishDuration;
     }
 
-    public byte getId(){
+    public int getId(){
         return this.id;
     }
